moved "api" to ui folder, and renamed to ui. also refactored the utility functions to utils
